refs #187 -- move Brazil tests to a standalone test module
